Add files via upload
AD-HOC Analysis & Insights :

Animals and science are the two most popular categories of content, showing that people enjoy "real life" and "factual" content the most.

Food is a common theme with the top 5 categories with "Health Eating" ranking the highest. This may give an indication to the audience within your user base. You could use this insight to create a campaign and work with healthy eating brands to boost user engagement.